Move Paths to play.core.j namespace
